LayoutTests/http/tests/w3c/dom/nodes/Node-textContent-expected.txt


PASS For an empty Element, textContent should be the empty string 
PASS For an empty DocumentFragment, textContent should be the empty string 
FAIL Element with children NotSupportedError: DOM Exception 9
FAIL Element with descendants NotSupportedError: DOM Exception 9
FAIL DocumentFragment with children NotSupportedError: DOM Exception 9
FAIL DocumentFragment with descendants NotSupportedError: DOM Exception 9
PASS For an empty Text, textContent should be the empty string 
FAIL For an empty ProcessingInstruction, textContent should be the empty string NotSupportedError: DOM Exception 9
PASS For an empty Comment, textContent should be the empty string 
PASS For a Text with data, textContent should be that data 
FAIL For a ProcessingInstruction with data, textContent should be that data NotSupportedError: DOM Exception 9
PASS For a Comment with data, textContent should be that data 
PASS For Documents created by parser, textContent should be null 
PASS For Documents created by createDocument, textContent should be null 
PASS For Documents created by createHTMLDocument, textContent should be null 
PASS For DocumentType created by parser, textContent should be null 
PASS For DocumentType created by script, textContent should be null 
PASS Element without children set to null 
PASS Element with empty text node as child set to null 
FAIL Element with children set to null NotSupportedError: DOM Exception 9
FAIL Element with descendants set to null NotSupportedError: DOM Exception 9
PASS DocumentFragment without children set to null 
PASS DocumentFragment with empty text node as child set to null 
FAIL DocumentFragment with children set to null NotSupportedError: DOM Exception 9
FAIL DocumentFragment with descendants set to null NotSupportedError: DOM Exception 9
FAIL Element without children set to undefined assert_equals: expected "" but got "undefined"
FAIL Element with empty text node as child set to undefined assert_equals: expected "" but got "undefined"
FAIL Element with children set to undefined NotSupportedError: DOM Exception 9
FAIL Element with descendants set to undefined NotSupportedError: DOM Exception 9
FAIL DocumentFragment without children set to undefined assert_equals: expected "" but got "undefined"
FAIL DocumentFragment with empty text node as child set to undefined assert_equals: expected "" but got "undefined"
FAIL DocumentFragment with children set to undefined NotSupportedError: DOM Exception 9
FAIL DocumentFragment with descendants set to undefined NotSupportedError: DOM Exception 9
PASS Element without children set to "" 
PASS Element with empty text node as child set to "" 
FAIL Element with children set to "" NotSupportedError: DOM Exception 9
FAIL Element with descendants set to "" NotSupportedError: DOM Exception 9
PASS DocumentFragment without children set to "" 
PASS DocumentFragment with empty text node as child set to "" 
FAIL DocumentFragment with children set to "" NotSupportedError: DOM Exception 9
FAIL DocumentFragment with descendants set to "" NotSupportedError: DOM Exception 9
PASS Element without children set to 42 
PASS Element with empty text node as child set to 42 
FAIL Element with children set to 42 NotSupportedError: DOM Exception 9
FAIL Element with descendants set to 42 NotSupportedError: DOM Exception 9
PASS DocumentFragment without children set to 42 
PASS DocumentFragment with empty text node as child set to 42 
FAIL DocumentFragment with children set to 42 NotSupportedError: DOM Exception 9
FAIL DocumentFragment with descendants set to 42 NotSupportedError: DOM Exception 9
PASS Element without children set to "abc" 
PASS Element with empty text node as child set to "abc" 
FAIL Element with children set to "abc" NotSupportedError: DOM Exception 9
FAIL Element with descendants set to "abc" NotSupportedError: DOM Exception 9
PASS DocumentFragment without children set to "abc" 
PASS DocumentFragment with empty text node as child set to "abc" 
FAIL DocumentFragment with children set to "abc" NotSupportedError: DOM Exception 9
FAIL DocumentFragment with descendants set to "abc" NotSupportedError: DOM Exception 9
PASS Element without children set to "<b>xyz</b>" 
PASS Element with empty text node as child set to "<b>xyz</b>" 
FAIL Element with children set to "<b>xyz</b>" NotSupportedError: DOM Exception 9
FAIL Element with descendants set to "<b>xyz</b>" NotSupportedError: DOM Exception 9
PASS DocumentFragment without children set to "<b>xyz</b>" 
PASS DocumentFragment with empty text node as child set to "<b>xyz</b>" 
FAIL DocumentFragment with children set to "<b>xyz</b>" NotSupportedError: DOM Exception 9
FAIL DocumentFragment with descendants set to "<b>xyz</b>" NotSupportedError: DOM Exception 9
PASS Element without children set to "d\0e" 
PASS Element with empty text node as child set to "d\0e" 
FAIL Element with children set to "d\0e" NotSupportedError: DOM Exception 9
FAIL Element with descendants set to "d\0e" NotSupportedError: DOM Exception 9
PASS DocumentFragment without children set to "d\0e" 
PASS DocumentFragment with empty text node as child set to "d\0e" 
FAIL DocumentFragment with children set to "d\0e" NotSupportedError: DOM Exception 9
FAIL DocumentFragment with descendants set to "d\0e" NotSupportedError: DOM Exception 9
PASS For a Text, textContent should set the data 
FAIL For a ProcessingInstruction, textContent should set the data NotSupportedError: DOM Exception 9
PASS For a Comment, textContent should set the data 
PASS For Documents created by parser, setting textContent should do nothing 
PASS For Documents created by createDocument, setting textContent should do nothing 
PASS For Documents created by createHTMLDocument, setting textContent should do nothing 
PASS For DocumentType created by parser, setting textContent should do nothing 
PASS For DocumentType created by script, setting textContent should do nothing 
